# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "VIVO API",
    "description": "Documentação VIVO API"
  },
  "paths": {
    "/api/balance": {
      "post": {
        "summary": "Registra um valor na balance",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "description": "value",
            "required": false,
            "type": "string",
            "readOnly": true
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso ao registar um valor na balance",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "POST /balance success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": {
                      "id": 1,
                      "value": 10.00,
                      "created_at": "2023-10-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao registar um valor na balance",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "POST /balance error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Obter todos os valores registrado na balance",
        "responses": {
          "200": {
            "description": "Sucesso ao obter todos os valores registrados na balance",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "GET /balance success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": [
                      {
                        "id": 2,
                        "value": "20.00",
                        "created_at": "2023-10-02T13:30:07.000Z",
                        "updated_at": null
                      },
                      {
                        "id": 3,
                        "value": "20.00",
                        "created_at": "2023-10-02T13:30:08.000Z",
                        "updated_at": null
                      },
                      {
                        "id": 4,
                        "value": "20.00",
                        "created_at": "2023-10-02T13:30:11.000Z",
                        "updated_at": null
                      }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Erro ao obter todos os valores registados na balance",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "GET /balance error"
                }
              }
            }
          }
        }
      }
    },
    "/api/balance/:id": {
      "get": {
        "summary": "Obtem o valor registrado na balance pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "type": "string",
            "format": "ObjectId",
            "readOnly": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successo ao obter um valor na balance pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "GET /balance/:id success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": {
                      "id": 2,
                      "value": 10,
                      "created_at": "2023-10-02T13:30:07.000Z",
                      "updated_at": "2023-10-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao obter um valor na balance pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "GET /balance/:id error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Editar valores na balance pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "type": "string",
            "format": "ObjectId",
            "readOnly": false
          },
          {
            "name": "value",
            "in": "body",
            "description": "value",
            "required": true,
            "type": "string",
            "readOnly": false
          },
          {
            "name": "updaetd_at",
            "in": "body",
            "description": "created_at EXP: 2023-10-01T00:00:00.000Z",
            "required": false,
            "type": "string",
            "readOnly": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successo ao um valor na balance pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "PUT /balance/:id success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": {
                      "id": 2,
                      "value": 10,
                      "created_at": "2023-10-02T13:30:07.000Z",
                      "updated_at": "2023-10-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao um valor na balance pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "PUT /balance/:id error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletar um valor na balance pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string",
            "format": "ObjectId",
            "readOnly": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successo ao deletar item da lista pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "DELETE /balance/:id success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": {
                      "id": 2,
                      "value": "10.00",
                      "created_at": "2023-10-02T13:30:07.000Z",
                      "updated_at": "2023-10-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao deletar um valor na balance pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "DELETE /balance/:id error"
                }
              }
            }
          }
        }
      }
    },
    "/api/list": {
      "post": {
        "summary": "Registra itens na lista",
        "parameters": [
          {
            "name": "description",
            "in": "body",
            "description": "description",
            "required": false,
            "type": "string",
            "readOnly": true
          },
          {
            "name": "value",
            "in": "body",
            "description": "value",
            "required": false,
            "type": "string",
            "readOnly": true
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso ao registar item a lista",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "POST /list success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": {
                      "id": 5,
                      "description": "Mercado",
                      "value": 10,
                      "created_at": "2023-10-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao registar item a lista",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "POST /list error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Obter todos os itens da lista",
        "responses": {
          "200": {
            "description": "Sucesso ao obter todos os itens a lista",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "GET /list success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": [
                      {
                          "id": 1,
                          "description": "Mercado",
                          "value": "10.00",
                          "created_at": "2023-10-01"
                      },
                      {
                          "id": 2,
                          "description": "Mercado",
                          "value": "10.00",
                          "created_at": "2023-10-01"
                      },
                      {
                          "id": 4,
                          "description": "Mercado",
                          "value": "10.00",
                          "created_at": "2023-10-01"
                      },
                      {
                          "id": 5,
                          "description": "Mercado",
                          "value": "10.00",
                          "created_at": "2023-10-01"
                      }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Erro ao obter todos os itens a lista",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "GET /list error"
                }
              }
            }
          }
        }
      }
    },
    "/api/list/:id": {
      "get": {
        "summary": "Obtem o item registrado na list pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "type": "string",
            "format": "ObjectId",
            "readOnly": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successo ao obter o item registrado na list pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "GET /list/:id success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": {
                      "id": 1,
                      "description": "Fármacia",
                      "value": "20.00",
                      "created_at": "2023-10-02T13:39:49.000Z",
                      "updated_at": null
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao obter o item registrado na list pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "GET /list/:id error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Editar itens na lista pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "type": "string",
            "format": "ObjectId",
            "readOnly": false
          },
          {
            "name": "description",
            "in": "body",
            "description": "description",
            "required": true,
            "type": "string",
            "readOnly": false
          },
          {
            "name": "value",
            "in": "body",
            "description": "value",
            "required": true,
            "type": "string",
            "readOnly": false
          },
          {
            "name": "updaetd_at",
            "in": "body",
            "description": "created_at EXP: 2023-10-01T00:00:00.000Z",
            "required": false,
            "type": "string",
            "readOnly": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successo ao editar item da lista pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "PUT /list/:id success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": {
                      "id": 1,
                      "description": "Farmácia",
                      "value": 10,
                      "created_at": "2023-10-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao editar item da lista pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "PUT /list/:id error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletar item da lista pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string",
            "format": "ObjectId",
            "readOnly": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successo ao deletar item da lista pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de sucesso",
                  "example": "DELETE /list/:id success"
                },
                "Success": {
                  "type": "integer",
                  "example": true
                },
                "Results": {
                    "type": "integer",
                    "example": {
                      "id": 3,
                      "description": "Farmácia",
                      "value": "10.00",
                      "created_at": "2023-10-01"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao deletar item da lista pelo ID",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string",
                  "description": "Mensagem de erro",
                  "example": "DELETE /list/:id error"
                }
              }
            }
          }
        }
      }
    }
  }
}